requirement.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11.11
django-paypal==0.5.0
idna==2.8
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
